Remove elkstack circular dep from platformstack
Due to the various problems it presents (circular dependencies, kibana Berksfile entries on every cookbook in a 3 mile radius, etc), we're no longer going to activate elkstack by default in platformstack. In order to include elkstack now, simply `include_recipe 'elkstack::agent'` on your node or add it to the runlist directly.